trofimov_hw_1_v2 (completed)
