+++ STUDENTS/Trofimov/trofimov_hw_1/src/index.html
@@ -30,7 +30,7 @@
                     <br>
                     <label for="player" class="text_number">Введите число -></label>
                     <br>
-                    <input type="button" id="gameProcess" name="ok" value="Готов ответить" onclick="gameProcess()" class="button_form color_changer">
+                    <input type="button" id="gameProcess" name="ok" value="Готов ответить" class="button_form color_changer">
                     <a href="index.html">
                         <input name="search1" type="button" class="knopka button_form color_changer" title="Обновить" value="Играть заново">
                     </a>
@@ -47,7 +47,5 @@
 
 
     </div>
-
-    <script src="./script.js"></script>
 </body>
 </html>
